requirements.txt
pybit==5.3.0
requests==2.31.0
websocket-client==1.5.2
